# tests/hosts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AssistantClarification": {
        "additionalProperties": false,
        "properties": {
          "choices": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Choices",
            "type": "array"
          },
          "kind": {
            "const": "clarification",
            "default": "clarification",
            "title": "Kind",
            "type": "string"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AssistantClarification",
        "type": "object"
      },
...
      "AssistantMessage": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "message",
            "default": "message",
            "title": "Kind",
            "type": "string"
          },
          "text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AssistantMessage",
        "type": "object"
      },
      "AssistantPendingAction": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PendingActionSummary"
          },
          "kind": {
            "const": "pending_action",
            "default": "pending_action",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "AssistantPendingAction",
        "type": "object"
      },
      "AssistantRecommendationPresentation": {
        "additionalProperties": false,
        "properties": {
          "intro": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Intro",
            "type": "string"
          },
          "kind": {
            "const": "recommendations",
            "default": "recommendations",
            "title": "Kind",
            "type": "string"
          },
          "recommendation_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Recommendation Ids",
            "type": "array"
          }
        },
        "required": [
          "intro",
          "recommendation_ids"
        ],
        "title": "AssistantRecommendationPresentation",
        "type": "object"
      },
      "AssistantTurnRequest": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "pattern": "^conv_[0-9a-f]{32}$",
            "title": "Conversation Id",
            "type": "string"
          },
          "locale": {
            "default": "en-US",
            "maxLength": 32,
            "minLength": 2,
            "title": "Locale",
            "type": "string"
          },
          "text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "text"
        ],
        "title": "AssistantTurnRequest",
        "type": "object"
      },
      "AssistantTurnResponse": {
        "additionalProperties": false,
        "properties": {
          "output": {
            "discriminator": {
              "mapping": {
                "clarification": "#/components/schemas/AssistantClarification",
                "message": "#/components/schemas/AssistantMessage",
                "pending_action": "#/components/schemas/AssistantPendingAction",
                "recommendations": "#/components/schemas/AssistantRecommendationPresentation"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/AssistantMessage"
              },
              {
                "$ref": "#/components/schemas/AssistantRecommendationPresentation"
              },
              {
                "$ref": "#/components/schemas/AssistantClarification"
              },
              {
                "$ref": "#/components/schemas/AssistantPendingAction"
              }
            ],
            "title": "Output"
          }
        },
        "required": [
          "output"
        ],
        "title": "AssistantTurnResponse",
        "type": "object"
      },
...
      "Conversation": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "pattern": "^conv_[0-9a-f]{32}$",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "retention_days": {
            "default": 30,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          },
          "scope": {
            "$ref": "#/components/schemas/ConversationScope"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "scope",
          "created_at",
          "updated_at"
        ],
        "title": "Conversation",
        "type": "object"
      },
      "ConversationMessage": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 16000,
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "message_id": {
            "pattern": "^msg_[0-9a-f]{32}$",
            "title": "Message Id",
            "type": "string"
          },
          "pending_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingActionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "references": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "References",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/ConversationRole"
          },
          "tool_calls": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ToolCallSummary"
            },
            "maxItems": 16,
            "title": "Tool Calls",
            "type": "array"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TurnUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_correction": {
            "default": false,
            "title": "User Correction",
            "type": "boolean"
          }
        },
        "required": [
          "message_id",
          "role",
          "content",
          "created_at",
          "idempotency_key"
        ],
        "title": "ConversationMessage",
        "type": "object"
      },
      "ConversationResponse": {
        "additionalProperties": false,
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "messages": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ConversationMessage"
            },
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "conversation"
        ],
        "title": "ConversationResponse",
        "type": "object"
      },
      "ConversationRole": {
        "enum": [
          "user",
          "assistant",
          "tool"
        ],
        "title": "ConversationRole",
        "type": "string"
      },
      "ConversationScope": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "local_user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Local User Id",
            "type": "string"
          }
        },
        "required": [
          "local_user_id",
          "device_id"
        ],
        "title": "ConversationScope",
        "type": "object"
      },
      "ConversationSummary": {
        "additionalProperties": false,
        "properties": {
          "confirmed_facts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Confirmed Facts",
            "type": "array"
          },
          "model_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "references": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "References",
            "type": "array"
          },
          "text": {
            "maxLength": 8000,
            "title": "Text",
            "type": "string"
          },
          "unresolved_actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PendingActionSummary"
            },
            "maxItems": 25,
            "title": "Unresolved Actions",
            "type": "array"
          },
          "user_corrections": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "User Corrections",
            "type": "array"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "text",
          "model_id",
          "version"
        ],
        "title": "ConversationSummary",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "validation",
          "unauthorized",
          "forbidden",
          "unavailable_capability",
          "conflict",
          "rate_limit",
          "temporary_failure",
          "not_found",
          "method_not_allowed"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "maxLength": 500,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PendingActionSummary": {
        "additionalProperties": false,
        "properties": {
          "effect": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Effect",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "pending_action_id": {
            "pattern": "^pending_[0-9a-f]{32}$",
            "title": "Pending Action Id",
            "type": "string"
          }
        },
        "required": [
          "pending_action_id",
          "effect",
          "expires_at"
        ],
        "title": "PendingActionSummary",
        "type": "object"
      },
...
      "ToolCallSummary": {
        "additionalProperties": false,
        "properties": {
          "outcome": {
            "enum": [
              "succeeded",
              "failed",
              "pending"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "safe_summary": {
            "maxLength": 1000,
            "title": "Safe Summary",
            "type": "string"
          },
          "tool_name": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "tool_name",
          "outcome",
          "safe_summary"
        ],
        "title": "ToolCallSummary",
        "type": "object"
      },
...
      "TurnUsage": {
        "additionalProperties": false,
        "properties": {
          "input_tokens": {
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "minimum": 0.0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "request_count": {
            "minimum": 0.0,
            "title": "Request Count",
            "type": "integer"
          }
        },
        "required": [
          "request_count",
          "input_tokens",
          "output_tokens"
        ],
        "title": "TurnUsage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenBiliClaw API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/assistant/conversations/{conversation_id}": {
      "get": {
        "operationId": "conversation_v1_assistant_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "message_limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Message Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Device-ID",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Conversation",
        "tags": [
          "assistant"
        ]
      }
    },
    "/v1/assistant/turns": {
      "post": {
        "operationId": "turn_v1_assistant_turns_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Device-ID",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Device-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Turn",
        "tags": [
          "assistant"
        ]
      }
    },
...
  }
}